1、将所有金额改成Integer类型，使用分作为单位（同时修改数据库中对应金额字段类型） 2、将所有微信命名的文件规范名称 3、修改微信异步通知处理方法
待开发：
1、分账处理逻辑
 (1)非分账支付的支付订单无法调用分账接口
 (2)只有完成支付的订单可以调用分账接口
 (3)如果已经退款则无法调用分账
2、分账记录数据表
3、分账回退数据表

+++ pay/src/main/java/com/yp/pay/entity/aliandwx/entity/TradeRefundRecordDO.java
@@ -3,7 +3,6 @@ package com.yp.pay.entity.aliandwx.entity;
 import lombok.Data;
 
 import javax.persistence.Table;
-import java.math.BigDecimal;
 import java.util.Date;
 
 
@@ -23,7 +22,7 @@ public class TradeRefundRecordDO {
 
     private String channelOrderNo;
 
-    private BigDecimal orderAmount;
+    private Integer orderAmount;
 
     /**
      * 支付渠道编码(WX_PAY:微信支付 ALI_PAY:支付宝支付)
@@ -44,9 +43,9 @@ public class TradeRefundRecordDO {
 
     private Date refundSuccessTime;
 
-    private BigDecimal refundApplyAmount;
+    private Integer refundApplyAmount;
 
-    private BigDecimal successRefundAmount;
+    private Integer successRefundAmount;
 
     private String orderFrom;
 
@@ -54,7 +53,7 @@ public class TradeRefundRecordDO {
 
     private String remark;
 
-    private BigDecimal merCost;
+    private Integer merCost;
 
     private String createUser;
